pyproject.toml
[project]
name = "to-webp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow-heif>=0.22.0",
]
